Move unit test wrapper generation into a test utils module
All the schema tests use the test utils module
